pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-trial-env"
version = "1.0.0"
description = "OpenEnv clinical trial protocol review environment for RL agent training"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.7.1,<3.0.0",
    "openai>=1.30.1",
    "requests>=2.31.0",
    "gradio>=4.44.1",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
]

[tool.setuptools]
packages = ["clinical_trial_env", "clinical_trial_env.server"]

[tool.setuptools.package-dir]
clinical_trial_env = "."

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]
